Move flags to separate files
