+++ cmd/lukso/main.go
@@ -89,14 +89,14 @@ REPO: https://github.com/lukso-network/tools-lukso-cli
 				{
 					Name:            "execution",
 					Usage:           "Outputs selected execution client's logs, add the network flag, if not mainnet",
-					Flags:           flags.GethLogsFlags,
+					Flags:           flags.ExecutionLogsFlags,
 					Action:          commands.SelectNetworkFor(cmd.LogsLayer(configs.ExecutionLayer)),
 					HideHelpCommand: true,
 				},
 				{
 					Name:            "consensus",
 					Usage:           "Outputs selected consensus client's logs, add the network flag, if not mainnet",
-					Flags:           flags.PrysmLogsFlags,
+					Flags:           flags.ConsensusLogsFlags,
 					Action:          commands.SelectNetworkFor(cmd.LogsLayer(configs.ConsensusLayer)),
 					HideHelpCommand: true,
 				},
