+++ flags/stop.go
@@ -1,0 +1,9 @@
+package flags
+
+import "github.com/urfave/cli/v2"
+
+var StopFlags = []cli.Flag{
+	executionSelectedFlag,
+	consensusSelectedFlag,
+	validatorSelectedFlag,
+}
